requirements.txt
autopep8
django==2.2.10
pycodestyle
pytz==2020.1
sqlparse==0.3.1

Unipath
python-decouple
whitenoise==5.2.0
psycopg2==2.8.6
gunicorn==20.0.4
pandas
sqlalchemy
django-filter==2.3.0

django-storages==1.10.1
s3transfer==0.3.3
urllib3==1.25.10
six==1.15.0
python-dateutil==2.8.1
jmespath==0.10.0
docutils==0.15.2
djangorestframework==3.11.1
django-bootstrap-form==3.4
dj-database-url==0.5.0
asgiref==3.2.10
boto3==1.14.62
botocore==1.17.62